Minor performance tweaks for FPU-less systems.
Add some more temporary variables of other types so that we can do a
little more integer math in `main()` where appropriate.  I've also
changed some of the types of internal functions, too.

+++ src/aprs-wx.h
@@ -85,7 +85,7 @@ packetConstructor (APRSPacket* const p);
 extern
 #endif
 void
-rain (char* const pResult, const double precip);
+rain (char* const pResult, const float precip);
 
 /**
  * notNull() -- return !0 if the user specified a meaningful value.
